move profiler
