+++ src/examples/simple_zkapp.ts
@@ -11,7 +11,7 @@ import {
   Bool,
   PublicKey,
 } from 'o1js';
-import { getProfiler } from './profiler.js';
+import { getProfiler } from './utils/profiler.js';
 
 const doProofs = true;
