+++ src/examples/zkapps/dex/upgradability.ts
@@ -8,7 +8,7 @@ import {
 } from 'o1js';
 import { createDex, TokenContract, addresses, keys, tokenIds } from './dex.js';
 import { expect } from 'expect';
-import { getProfiler } from '../../profiler.js';
+import { getProfiler } from '../../utils/profiler.js';
 
 await isReady;
