+++ src/examples/zkapps/hello_world/run.ts
@@ -1,5 +1,5 @@
 import { AccountUpdate, Field, Mina, PrivateKey } from 'o1js';
-import { getProfiler } from '../../profiler.js';
+import { getProfiler } from '../../utils/profiler.js';
 import { HelloWorld, adminPrivateKey } from './hello_world.js';
 
 const HelloWorldProfier = getProfiler('Hello World');
