+++ src/examples/zkapps/voting/run.ts
@@ -8,7 +8,7 @@ import {
 import { OffchainStorage } from './off_chain_storage.js';
 import { Member } from './member.js';
 import { testSet } from './test.js';
-import { getProfiler } from '../../profiler.js';
+import { getProfiler } from '../../utils/profiler.js';
 
 console.log('Running Voting script...');
